+++ Graph.cpp
@@ -1,14 +1,14 @@
 #include "Graph.hpp"
 
-Graph::Graph(int vertices) : numVertices(vertices), adjList(vertices) {}
+Graph::Graph(size_t vertices) : numVertices(vertices), adjList(vertices) {}
 
-void Graph::addEdge(int src, int dest, int weight) {
+void Graph::addEdge(size_t src, size_t dest, size_t weight) {
     adjList[src].emplace_back(dest, weight);
 }
 
 const std::string Graph::getGraphAsString() const {
     std::ostringstream os;
-    for (int i = 0; i < numVertices; i++) {
+    for (size_t i = 0; i < numVertices; i++) {
         os << "Adjacency list of vertex " << i << ": ";
         for(const auto& edge : adjList[i]) {
             os << " -> (" << edge.destination << "," << edge.weight << ") ";
